+++ src/hittable/rotate.rs
@@ -8,29 +8,29 @@ use super::Hittable;
 
 pub struct RotateY {
     ptr: Box<dyn Hittable>,
-    sin_theta: f64,
-    cos_theta: f64,
+    sin_theta: f32,
+    cos_theta: f32,
     bbox: Option<Aabb>,
 }
 
 impl RotateY {
-    pub fn new(ptr: Box<dyn Hittable>, angle: f64) -> Self {
+    pub fn new(ptr: Box<dyn Hittable>, angle: f32) -> Self {
         let radians = angle.to_radians();
         let sin_theta = radians.sin();
         let cos_theta = radians.cos();
         let bbox = ptr.bounding_box(0.0, 1.0);
 
-        let mut min = Point3::new(f64::INFINITY, f64::INFINITY, f64::INFINITY);
-        let mut max = Point3::new(f64::NEG_INFINITY, f64::NEG_INFINITY, f64::NEG_INFINITY);
+        let mut min = Point3::new(f32::INFINITY, f32::INFINITY, f32::INFINITY);
+        let mut max = Point3::new(f32::NEG_INFINITY, f32::NEG_INFINITY, f32::NEG_INFINITY);
 
         let bbmax = bbox.as_ref().map(|bb| bb.max).unwrap_or(Vec3::zeros());
         let bbmin = bbox.as_ref().map(|bb| bb.min).unwrap_or(Vec3::zeros());
         for i in 0..2 {
             for j in 0..2 {
                 for k in 0..2 {
-                    let x = i as f64 * bbmax.x + (1 - i) as f64 * bbmin.x;
-                    let y = j as f64 * bbmax.y + (1 - j) as f64 * bbmin.y;
-                    let z = k as f64 * bbmax.z + (1 - k) as f64 * bbmin.z;
+                    let x = i as f32 * bbmax.x + (1 - i) as f32 * bbmin.x;
+                    let y = j as f32 * bbmax.y + (1 - j) as f32 * bbmin.y;
+                    let z = k as f32 * bbmax.z + (1 - k) as f32 * bbmin.z;
 
                     let newx = cos_theta * x + sin_theta * z;
                     let newz = -sin_theta * x + cos_theta * z;
@@ -55,7 +55,7 @@ impl RotateY {
 }
 
 impl Hittable for RotateY {
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let mut origin = ray.origin;
         let mut direction = ray.direction;
 
@@ -89,7 +89,7 @@ impl Hittable for RotateY {
         ))
     }
 
-    fn bounding_box(&self, _time0: f64, _time1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _time0: f32, _time1: f32) -> Option<Aabb> {
         self.bbox.clone()
     }
 }
