+++ src/ray.rs
@@ -3,11 +3,11 @@ use crate::Vec3;
 pub struct Ray {
     pub origin: Vec3,
     pub direction: Vec3,
-    pub time: f64,
+    pub time: f32,
 }
 
 impl Ray {
-    pub fn new(origin: Vec3, direction: Vec3, time: f64) -> Self {
+    pub fn new(origin: Vec3, direction: Vec3, time: f32) -> Self {
         Self {
             origin,
             direction,
@@ -15,7 +15,7 @@ impl Ray {
         }
     }
 
-    pub fn at(&self, t: f64) -> Vec3 {
+    pub fn at(&self, t: f32) -> Vec3 {
         self.origin + t * self.direction
     }
 }
